refactor: 구분자 관련 클래스 디렉토리 변경
