+++ src/main/java/calculator/model/delimiter/CustomDelimiterPattern.java
@@ -1,4 +1,4 @@
-package calculator.model;
+package calculator.model.delimiter;
 
 public enum CustomDelimiterPattern {
 	START("//"),
